Refactored sprite drawing code, added empty item classes, implemented AbstractItem class, added custom math class, added next/previous item commands, improved keyboard controller, added 2 default items to display
Refactored sprite drawing, some areas of code are redundant as a result and need to be fixed. Added empty items that will be implemented later. Implemented a Clock Item and Bow Item that use a default ball sprite, will be changed later. Added items to the main game that can be switched between. Added a custom math class to hold helper math functions, such as mathematical mod which is used to switch between items on screen.